+++ changelog.md
@@ -1,3 +1,11 @@
-# 0.1.0
+# 0.1.2
+- Added a placeholder for the mod: TAB by NEZNAMY
+- Added support for color coded faction names (only with vanilla color coding using &. Gradients and hex aren't supported yet.)
 
+# 0.1.1
+- Added a cooldown to the /faction home command
+- Added config values related to the cooldown for /faction home
+- Added a feature that prevents a faction from being knocked out when there is no player of said faction  online
+
+# 0.1.0
 First version: No documented changes